requirements.txt
pandas>=1.5.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=0.19.0
pathlib2>=2.3.0
